コマンドライン引数で指定された整数を半径とした円の面積を求めよ。ただし円周率は3.14とする ※再度自身にプルリクを送るために"to_1 -> to_f"の変更

+++ 2-2/main.rb
@@ -1,4 +1,4 @@
-radius = ARGV[0].to_i
+radius = ARGV[0].to_f
 pi     = 3.14
 
 puts "#{radius**2*pi}"
